requirements.txt
streamlit>=1.32
openai>=1.13
passlib[bcrypt]>=1.7
python-dotenv>=1.0
